backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
PyPDF2==3.0.1
python-docx==1.1.2
sentence-transformers==3.1.1
chromadb==0.5.7
httpx==0.27.2

# Auth & Security
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
cryptography==43.0.0
email-validator==2.2.0

# MongoDB
motor==3.6.0
pymongo==4.9.1

# LLM Providers
anthropic==0.39.0
openai==1.51.0

# File handling
python-magic==0.4.27
aiofiles==24.1.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
